reorganized create service methods to predix.admin.app, add_to_manifest now takes manifest obj instead of path
