+++ predix/admin/ie/parking.py
@@ -1,7 +1,6 @@
 
 import os
 
-import predix.app
 import predix.security.uaa
 import predix.admin.service
 
@@ -58,13 +57,15 @@ class ParkingPlanning(object):
         """
         return self.service.settings.data['zone']['http-header-value']
 
-    def add_to_manifest(self, manifest_path):
+    def add_to_manifest(self, manifest):
         """
-        Add details to the manifest that applications using
-        this service may need to consume.
-        """
-        manifest = predix.app.Manifest(manifest_path)
+        Add useful details to the manifest about this service
+        so that it can be used in an application.
 
+        :param manifest: An predix.admin.app.Manifest object
+            instance that manages reading/writing manifest config
+            for a cloud foundry app.
+        """
         # Add this service to list of services
         manifest.add_service(self.service.name)
