+++ predix/admin/weather.py
@@ -1,7 +1,6 @@
 
 import os
 
-import predix.app
 import predix.config
 import predix.admin.service
 import predix.data.weather
@@ -38,9 +37,15 @@ class WeatherForecast(object):
         uri = predix.config.get_env_key(self.use_class, 'uri')
         os.environ[uri] = self.service.settings.data['uri']
 
-    def add_to_manifest(self, manifest_path):
-        manifest = predix.app.Manifest(manifest_path)
+    def add_to_manifest(self, manifest):
+        """
+        Add useful details to the manifest about this service
+        so that it can be used in an application.
 
+        :param manifest: An predix.admin.app.Manifest object
+            instance that manages reading/writing manifest config
+            for a cloud foundry app.
+        """
         # Add this service to list of services
         manifest.add_service(self.service.name)
