+++ predix/app.py
@@ -11,7 +11,7 @@ class Manifest(object):
     run our applications the manifest is a place to store
     important configuration details.
     """
-    def __init__(self, manifest_path, app_name='my-predix-app', debug=False):
+    def __init__(self, manifest_path='manifest.yml', app_name='my-predix-app', debug=False):
         self.manifest_path = os.path.expanduser(manifest_path)
         self.app_name = app_name
 
@@ -56,45 +56,6 @@ class Manifest(object):
 
         self.write_manifest()
 
-    def create_manifest_from_space(self):
-        """
-        Populate a manifest file generated from details from the
-        cloud foundry space environment.
-        """
-        import predix.admin.cf.spaces
-        space = predix.admin.cf.spaces.Space()
-
-        summary = space.get_space_summary()
-        for instance in summary['services']:
-            service_type = instance['service_plan']['service']['label']
-            name = instance['name']
-            if service_type == 'predix-uaa':
-                import predix.admin.uaa
-                uaa = predix.admin.uaa.UserAccountAuthentication(name=name)
-                uaa.add_to_manifest(self.manifest_path)
-            elif service_type == 'predix-acs':
-                import predix.admin.acs
-                acs = predix.admin.acs.AccessControl(name=name)
-                acs.add_to_manifest(self.manifest_path)
-            elif service_type == 'predix-asset':
-                import predix.admin.asset
-                asset = predix.admin.asset.Asset(name=name)
-                asset.add_to_manifest(self.manifest_path)
-            elif service_type == 'predix-timeseries':
-                import predix.admin.timeseries
-                timeseries = predix.admin.timeseries.TimeSeries(name=name)
-                timeseries.add_to_manifest(self.manifest_path)
-            elif service_type == 'predix-blobstore':
-                import predix.admin.blobstore
-                blobstore = predix.admin.blobstore.BlobStore(name=name)
-                blobstore.add_to_manifest(self.manifest_path)
-            elif service_type == 'us-weather-forecast':
-                import predix.admin.weather
-                weather = predix.admin.weather.WeatherForecast(name=name)
-                weather.add_to_manifest(self.manifest_path)
-            else:
-                logging.warn("Unsupported service type: %s" % service_type)
-
     def write_manifest(self):
         """
         Write manifest to disk.
@@ -110,6 +71,7 @@ class Manifest(object):
         variable.
         """
         self.manifest['env'][key] = value
+        os.environ[key] = value
 
     def add_service(self, service_name):
         """
@@ -136,7 +98,7 @@ class Manifest(object):
         needed scopes and authorities for the services
         in this manifest.
         """
-        key = 'PREDIX_SECURITY_UAA_CLIENT_ID'
+        key = 'PREDIX_APP_CLIENT_ID'
         if key not in self.manifest['env']:
             raise ValueError("%s undefined in manifest." % key)
 
@@ -148,27 +110,12 @@ class Manifest(object):
         Return the client secret that should correspond with
         the client id.
         """
-        key = 'PREDIX_SECURITY_UAA_CLIENT_SECRET'
+        key = 'PREDIX_APP_CLIENT_SECRET'
         if key not in self.manifest['env']:
             raise ValueError("%s must be added to manifest." % key)
 
         self.client_secret = self.manifest['env'][key]
         return self.client_secret
-
-    def create_timeseries(self):
-        """
-        Creates an instance of the Time Series Service.
-        """
-        import predix.admin.timeseries
-        ts = predix.admin.timeseries.TimeSeries()
-        ts.create()
-
-        client_id = self.get_client_id()
-        if client_id:
-            ts.grant_client(client_id)
-
-        ts.add_to_manifest(self.manifest_path)
-        return ts
 
     def get_timeseries(self, *args, **kwargs):
         """
@@ -178,21 +125,6 @@ class Manifest(object):
         ts = predix.data.timeseries.TimeSeries(*args, **kwargs)
         return ts
 
-    def create_asset(self):
-        """
-        Creates an instance of the Asset Service.
-        """
-        import predix.admin.asset
-        asset = predix.admin.asset.Asset()
-        asset.create()
-
-        client_id = self.get_client_id()
-        if client_id:
-            asset.grant_client(client_id)
-
-        asset.add_to_manifest(self.manifest_path)
-        return asset
-
     def get_asset(self):
         """
         Returns an instance of the Asset Service.
@@ -200,27 +132,6 @@ class Manifest(object):
         import predix.data.asset
         asset = predix.data.asset.Asset()
         return asset
-
-    def create_uaa(self, admin_secret):
-        """
-        Creates an instance of UAA Service.
-        """
-        import predix.admin.uaa
-        uaa = predix.admin.uaa.UserAccountAuthentication()
-        if not uaa.exists():
-            uaa.create(admin_secret)
-            uaa.add_to_manifest(self.manifest_path)
-        return uaa
-
-    def create_client(self, client_id, client_secret):
-        """
-        Create a client and add it to the manifest.
-        """
-        import predix.admin.uaa
-        uaa = predix.admin.uaa.UserAccountAuthentication()
-        uaa.create_client(client_id, client_secret)
-        uaa.add_client_to_manifest(client_id, client_secret,
-                self.manifest_path)
 
     def get_uaa(self):
         """
@@ -230,22 +141,6 @@ class Manifest(object):
         uaa = predix.security.uaa.UserAccountAuthentication()
         return uaa
 
-    def create_acs(self):
-        """
-        Creates an instance of the Asset Service.
-        """
-        import predix.admin.acs
-        acs = predix.admin.acs.AccessControl()
-        acs.create()
-
-        client_id = self.get_client_id()
-        if client_id:
-            acs.grant_client(client_id)
-
-        acs.grant_client(client_id)
-        acs.add_to_manifest(self.manifest_path)
-        return acs
-
     def get_acs(self):
         """
         Returns an instance of the Asset Control Service.
@@ -254,22 +149,6 @@ class Manifest(object):
         acs = predix.security.acs.AccessControl()
         return acs
 
-    def create_weather(self):
-        """
-        Creates an instance of the Asset Service.
-        """
-        import predix.admin.weather
-        weather = predix.admin.weather.WeatherForecast()
-        weather.create()
-
-        client_id = self.get_client_id()
-        if client_id:
-            weather.grant_client(client_id)
-
-        weather.grant_client(client_id)
-        weather.add_to_manifest(self.manifest_path)
-        return weather
-
     def get_weather(self):
         """
         Returns an instance of the Weather Service.
@@ -277,17 +156,6 @@ class Manifest(object):
         import predix.data.weather
         weather = predix.data.weather.WeatherForecast()
         return weather
-
-    def create_blobstore(self):
-        """
-        Creates an instance of the BlobStore Service.
-        """
-        import predix.admin.blobstore
-        blobstore = predix.admin.blobstore.BlobStore()
-        blobstore.create()
-
-        blobstore.add_to_manifest(self.manifest_path)
-        return blobstore
 
     def get_blobstore(self):
         import predix.data.blobstore
